Refactor client common methods
